add private interface to configs with jgroups socket bindings, on all supported server migrations
