+++ eap6-to-eap7/src/main/java/org/jboss/migration/eap6/to/eap7/tasks/EnableHttpInterfaceSupportForHttpUpgrade.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.jboss.migration.eap6.to.eap7.standalone;
+package org.jboss.migration.eap6.to.eap7.tasks;
 
 import org.jboss.as.controller.PathAddress;
 import org.jboss.as.controller.operations.common.Util;
@@ -25,37 +25,26 @@ import org.jboss.migration.core.ServerMigrationTaskName;
 import org.jboss.migration.core.ServerMigrationTaskResult;
 import org.jboss.migration.core.ServerMigrationTasks;
 import org.jboss.migration.core.ServerPath;
-import org.jboss.migration.core.env.MigrationEnvironment;
 import org.jboss.migration.eap.EAP6Server;
-import org.jboss.migration.wfly10.config.management.SocketBindingsManagement;
-import org.jboss.migration.wfly10.config.task.SocketBindingsMigration;
+import org.jboss.migration.wfly10.config.management.ManagementInterfacesManagement;
+import org.jboss.migration.wfly10.config.task.ManagementInterfacesMigration;
 
 import static org.jboss.as.controller.descriptions.ModelDescriptionConstants.NAME;
 import static org.jboss.as.controller.descriptions.ModelDescriptionConstants.VALUE;
 import static org.jboss.as.controller.descriptions.ModelDescriptionConstants.WRITE_ATTRIBUTE_OPERATION;
 
 /**
- * Updates the management-https socket binding config.
+ * Updates the management interface http-interface to support http upgrade.
  * @author emmartins
  */
-public class UpdateManagementHttpsSocketBinding implements SocketBindingsMigration.SubtaskFactory<ServerPath<EAP6Server>> {
+public class EnableHttpInterfaceSupportForHttpUpgrade implements ManagementInterfacesMigration.SubtaskFactory<ServerPath<EAP6Server>> {
 
-    public static final String SERVER_MIGRATION_TASK_NAME_NAME = "update-management-https";
+    public static final String SERVER_MIGRATION_TASK_NAME_NAME = "enable-http-upgrade-support";
     public static final ServerMigrationTaskName SERVER_MIGRATION_TASK_NAME = new ServerMigrationTaskName.Builder(SERVER_MIGRATION_TASK_NAME_NAME).build();
-
-    public interface EnvironmentProperties {
-        /**
-         * the prefix for the name of the management-https socket binding related properties
-         */
-        String PROPERTIES_PREFIX = SocketBindingsMigration.SOCKET_BINDINGS + "." + SERVER_MIGRATION_TASK_NAME_NAME + ".";
-
-        String PORT = PROPERTIES_PREFIX + "port";
-    }
-
-    public static final String DEFAULT_PORT = "${jboss.management.https.port:9993}";
+    private static final String INTERFACE_NAME = "http-interface";
 
     @Override
-    public void addSubtasks(ServerPath<EAP6Server> source, final SocketBindingsManagement resourceManagement, ServerMigrationTasks subtasks) throws Exception {
+    public void addSubtasks(ServerPath<EAP6Server> source, final ManagementInterfacesManagement resourceManagement, ServerMigrationTasks subtasks) throws Exception {
         final ServerMigrationTask subtask = new ServerMigrationTask() {
             @Override
             public ServerMigrationTaskName getName() {
@@ -63,21 +52,17 @@ public class UpdateManagementHttpsSocketBinding implements SocketBindingsMigrati
             }
             @Override
             public ServerMigrationTaskResult run(ServerMigrationTaskContext context) throws Exception {
-                final MigrationEnvironment env = context.getServerMigrationContext().getMigrationEnvironment();
-                String envPropertyPort = env.getPropertyAsString(EnvironmentProperties.PORT);
-                if (envPropertyPort == null || envPropertyPort.isEmpty()) {
-                    envPropertyPort = DEFAULT_PORT;
-                }
-                if (!resourceManagement.getResourceNames().contains("management-https")) {
+                if (!resourceManagement.getResourceNames().contains(INTERFACE_NAME)) {
                     return ServerMigrationTaskResult.SKIPPED;
                 }
-                // management-https binding found, update port
-                final PathAddress pathAddress = resourceManagement.getResourcePathAddress("management-https");
+                // http interface found, turn on http upgrade
+                final PathAddress pathAddress = resourceManagement.getResourcePathAddress(INTERFACE_NAME);
+                //pathAddress(pathElement(CORE_SERVICE, MANAGEMENT), pathElement(MANAGEMENT_INTERFACE, "http-interface"));
                 final ModelNode writeAttrOp = Util.createEmptyOperation(WRITE_ATTRIBUTE_OPERATION, pathAddress);
-                writeAttrOp.get(NAME).set("port");
-                writeAttrOp.get(VALUE).set(envPropertyPort);
+                writeAttrOp.get(NAME).set("http-upgrade-enabled");
+                writeAttrOp.get(VALUE).set(true);
                 resourceManagement.getServerConfiguration().executeManagementOperation(writeAttrOp);
-                context.getLogger().infof("Socket binding 'management-https' default port set to "+envPropertyPort+".");
+                context.getLogger().infof("Activated HTTP Management Interface's support for HTTP Upgrade.");
                 return ServerMigrationTaskResult.SUCCESS;
             }
         };
